Improve jumps by specifying the cursor’s position.
By using \zs or \ze it is possible to place the cursor at the desired
position after a jump. So, for example, when jumping to a topic that is
prefixed by a checkbox, the cursor jumps directly into the checkbox.

This commit implements this feature with \zs. Using \ze (combined with
the e flag from search()) might be better, because \zs needs to be
removed from the regex in some cases (see topic_search()). In fact,
consider this line:

1  === 1
2  === 2
3    === X

With the cursor anywhere on line 3, the following regex always returns
3, as expected (or, at least, as we wish):

echo search('^\s*=== \S', 'bcnW')

(the c flag is needed to return 3 when the cursor is at the start of
line 3). The above regex is read as: starting from the cursor’s position,
search for a line that starts with zero or more spaces followed by ===,
another space and a non-space character.

The following search, however, returns 2 if the cursor is before X on
line 3:

echo search('^\s*=== \zs\S', 'bcnW')

This regex is read as follows: starting from the cursor’s position,
search for a non-space character preceded by a space, ===, and possibly
more space till the start of the line. If the cursor is on a = sign on
line 3, that character matches \S, but the regex fails to match because
that = is not preceded by ^\s*===. So, the first match is found in the
preceding line.

With \ze, the problem above is avoided. Consider:

echo search('^\s*=== \ze\S', 'bcnW')

This is read as: starting from the cursor position, search for a line
that starts with zero or more spaces followed by === and one more space,
which is also followed by a non-space character. This regex matches on
line 3 no matter where the cursor is placed on line 3.