Add g:outlaw_body_text_level.
Choose between two different modes of folding the body text. When the
variable is set to =, the body text has the same fold level has the
preceding topic header, hence it is folded along with it. When the
variable has a high numeric value (typically, 20), the body text is
folded independently. The latter also allows a fully expanded outline
with collapsed notes.

Both views are useful, so I am planning to add a toggle to switch
between the two.