Add prototype indentation function.
This is not working correctly yet, but it gives the idea. Re-indenting
an outline requires to keep some state (g:shifts), and the tricky part
is when/how to initialize such state, given that the indentexpr
mechanism does not provide such an initialization hook.

To see why some state is necessary, consider the following example:

=== h1
          === h1.1
    === h1.2

We assume that the correct indentation (assuming a two-space
shiftwidth) should be:

=== h1
  === h1.1
  === h1.2

(Note that keeping h1.2 as an uncle of h1.1 is not possible, since h1.2
would then have negative indentation).

Proceeding from top to bottom, we see that:

1. h1 is indented correctly, nothing to be done.
2. h1.1 is too far to the right, and should be moved below h1:

=== h1
  === h1.1
    === h1.2

3. When it comes to h1.2, we note that the current layout is in
principle correct. However, h1.2 was originally a child of h1, not of
h1.1. Remembering that, we know that it too must be shifted to the left,
resulting in the desired outline:

=== h1
  === h1.1
  === h1.2

Finally, note that the indentation function never moves topics to the
right: it either leaves then where they are or shift them to the left.

For now, I'm not taking care of notes (their indentation is never
changed, but in some cases it should).