Add new option: g:outlaw_indent.
If 'shiftwidth' is not set in a user's vimrc, the syntax file may use a
value of 'shiftwidth' that is subsequently updated, e.g., by a
document's modeline. This new option helps avoid this issue.

The user must set g:outlaw_indent in the user's vimrc when 'shiftwidth'
is not set in the vimrc or when the user wants to use a different value
in Outlaw buffers.

Fixes https://github.com/lifepillar/vim-outlaw/issues/1.