Make topic search patterns more robust.
Make sure that the topic mark is matched at the beginning of a line,
i.e., that it is preceded only by whitespace.

This change makes it possible to define arbitrary topic patterns. An
application of this will come with the next commit, where support for
the TaskPaper format is added very easily.