Add support for the TaskPaper format. 🎉
The (extremely simple) TaskPaper format is described here:

    https://guide.taskpaper.com/getting-started/

Syntax highlighting for tags had already been added. This commit
adds filetype detection (files with .taskpaper suffix) and redefines
the topic mark (locally) to match TaskPaper projects and tasks.

The regular expression matches either a line starting with - followed by
a space, or any non-whitespace character followed by anything ending
with a colon. If there is another non-whitespace character after the
colon, it must be the start of a tag.

The second branch is wrapped inside \%(\)\@=, so that the whole project
line matches with zero width. This is so that OutlawAddSibling() won’t
copy the current project text when adding a new topic. Note that \ze
cannot be used for this purpose, because \zs and \ze are stripped by
OutlawTopicPattern() (invoked by OutlawAddSibling()).
Another consequence is that OutlawFoldText() will retain the whole
project line when the topic is folded.

Note that TaskPaper documents are indented using tabs, so we unset
expandtab, too.